requirements.txt
amqp==5.0.7
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-extensions==3.1.5
django-timezone-field==4.2.1
djangorestframework==3.12.4
Faker==10.0.0
flower==1.0.0
humanize==3.13.1
ipython==7.30.1
jedi==0.18.1
kombu==5.2.2
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
traitlets==5.1.1
vine==5.0.0
wcwidth==0.2.5
